+++ app/Filters/AlreadyLoggedInFilter.php
@@ -9,7 +9,7 @@ use CodeIgniter\Filters\FilterInterface;
 class AlreadyLoggedInFilter implements FilterInterface
 {
     public function before(RequestInterface $request, $arguments = null)
-    {
+    {        
         if(session()->has('loggedUser')){
             return redirect()->back();
         }