+++ docs/changelog.md
@@ -1,3 +1,6 @@
+##### v0.0.18
+- Fixed bug in Config/Routes
+- Create message in View admin/login
 ##### v0.0.17
 - Fixed bug in Config/Routes
 - Create Filter AlreadyLoggedInFilter